Move Pulse publisher classes into Pulse service
