+++ treeherder/services/pulse/__init__.py
@@ -2,12 +2,14 @@ from .connection import pulse_conn
 from .consumers import (JobConsumer,
                         PushConsumer)
 from .exchange import get_exchange
+from .publisher import TreeherderPublisher
 from .sources import (job_sources,
                       push_sources)
 
 __all__ = [
     "JobConsumer",
     "PushConsumer",
+    "TreeherderPublisher",
     "get_exchange",
     "job_sources",
     "pulse_conn",
